refactor: isolate historical RunState compaction restoration
