+++ .changeset/quiet-legacy-compaction.md
@@ -1,0 +1,5 @@
+---
+'@openai/agents-core': patch
+---
+
+refactor: isolate historical RunState compaction restoration without changing snapshot compatibility or session reconciliation.
